+++ package.json
@@ -14,7 +14,7 @@
     "write-heading-ids": "docusaurus write-heading-ids",
     "lint": "stylelint '**/*.css' --fix --ignore-pattern build/**/*.css",
     "update-news": "./bin/update-news",
-    "backdrop": "cd bin/backdrop && npm install --silent && npm run -- backdrop"
+    "banner": "cd bin/banner && npm install --silent && npm run -- banner"
   },
   "dependencies": {
     "@docusaurus/core": "2.3.1",
